# Optimisation for /v3/courses SQL - remove double join to `course_site`