--- a/app/models/course.rb
+++ b/app/models/course.rb
@@ -144,11 +144,11 @@ class Course < ApplicationRecord
           class_name: "CourseEnrichment"
 
   scope :within, ->(range, origin:) do
-    joins(site_statuses: :site).merge(SiteStatus.where(status: %i[new_status running])).merge(Site.within(range, origin: origin))
+    joins(site_statuses: :site).merge(SiteStatus.where(status: :running)).merge(Site.within(range, origin: origin))
   end
 
   scope :by_distance, ->(origin:) do
-    joins(site_statuses: :site).merge(SiteStatus.where(status: %i[new_status running])).merge(Site.by_distance(origin: origin))
+    joins(site_statuses: :site).merge(SiteStatus.where(status: :running)).merge(Site.by_distance(origin: origin))
   end
 
   scope :by_name_ascending, -> do