@stevehook 2021-09-29
> This is looking great.
> 
> > How can we mitigate the risk?
> 
> You can use https://github.com/DFE-Digital/teacher-training-api/blob/master/spec/services/course_search_service_spec.rb#L653-L758 as a basis for the full integration tests in order to refactor https://github.com/DFE-Digital/teacher-training-api/blob/master/spec/services/course_search_service_spec.rb#L175-L199 to remove mockery and to assert confidence and correctiveness of behaviour.

We've logged a separate card to create integration specs for this service. Not sure how long it will take.