@davidgisbey 2021-09-30
We've got a separate card to rework the spec here https://trello.com/c/ugmvkcYf/4000-rewrite-the-tests-for-ttapi-coursesearchservice